+++ Arex388.NhtsaVpic/Models/ResponseBase`T.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Arex388.NhtsaVpic;
+
+public class ResponseBase<T> :
+    ResponseBase {
+    public IEnumerable<T> Results { get; set; } = Enumerable.Empty<T>();
+}